docker-mysql-config

+++ products/views.py
@@ -11,7 +11,6 @@ from django.contrib.auth.models import User
 class get_post_products(APIView):
     queryset = Products.objects.all()
     model = Products
-    # permission_classes=[IsAuthenticated]
 
     def get(self, request):
         params = self.request.query_params.get('search', None)
@@ -25,15 +24,17 @@ class get_post_products(APIView):
 
     def post(self, request, format=None):
         product_ser = Products_serializer(data=request.data)
-        if product_ser.is_valid():
-            user = User.objects.get(username=request.data['prd_owner'])
-            product_ser.save(prd_owner=user)
-            return Response(product_ser.data, status=status.HTTP_201_CREATED)
-        return Response(product_ser.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated:
+            if product_ser.is_valid():
+                product_ser.save(prd_owner=User.objects.get(username=request.user.username))
+                return Response(product_ser.data, status=status.HTTP_201_CREATED)
+            return Response(product_ser.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response({"error":"you have to login to add a product"})    
 
 
 class get_update_delete_product(APIView):
-    # permission_classes=[IsAuthenticated]
+    permission_classes = [IsAuthenticated]
+
     def get_object(self, pk):
         try:
             return Products.objects.get(pk=pk)
@@ -49,11 +50,15 @@ class get_update_delete_product(APIView):
         product = self.get_object(pk)
         products_ser = Products_serializer(product, data=request.data)
         if products_ser.is_valid():
-            products_ser.save()
-            return Response(products_ser.data)
+            if product.prd_owner == request.user:
+                products_ser.save()
+                return Response(products_ser.data)
+            return Response({"error": "only prodcut owner can modify "})
         return Response(products_ser.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk, format=None):
         product = self.get_object(pk)
-        product.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if product.prd_owner == request.user:
+            product.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response({"error": "only prodct owner can delete"})
